Relese candi

+++ app/Http/Controllers/Api/OperadorController.php
@@ -7,6 +7,7 @@ use App\Models\Operador;
 use App\Http\Requests\StoreOperadorRequest;
 use App\Http\Requests\UpdateOperadorRequest;
 use App\Http\Resources\OperadorResource;
+use App\Models\User;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request as HttpRequest;
@@ -19,10 +20,39 @@ class OperadorController extends Controller
      */
     public function index()
     {
-        return response( OperadorResource::collection(
+        $this->authorize('viewAny', Operador::class);
+        return response(OperadorResource::collection(
             Operador::all()
-        ),200);
+        ), 200);
         //
+    }
+
+    public function store_2(StoreOperadorRequest $request)
+    {
+        try {
+            $data = $request->validated();
+            $user = new User();
+            $user->name = $data["name"];
+            $user->email = $data["email"];
+            $user->password = bcrypt($data["password"]);
+            $user->save();
+
+            $operador = new Operador();
+            $operador->id_user = $user->id;
+            $operador->codigo_empleado = $data["codigo_empleado"];
+            $operador->nombre = $data["nombre"];
+            $operador->apellido_paterno = $data["apellido_paterno"];
+            $operador->apellido_materno = $data["apellido_materno"];
+            $operador->CURP = $data["CURP"];
+            $operador->fecha_nacimiento = $data["fecha_nacimiento"];
+            $operador->save();
+
+            return response(new OperadorResource($operador), 201);
+        } catch (Exception $e) {
+            return response()->json([
+                'error' => 'No se pudo guardar el operador'
+            ], 500);
+        }
     }
 
     /**
@@ -30,44 +60,44 @@ class OperadorController extends Controller
      */
     public function store(Operador $operador, StoreOperadorRequest $request)
     {
-        
-        try{
+        $this->authorize('create', Operador::class);
+
+        try {
             //VALIDA EL STORE
-        $data = $request->validated();
+            $data = $request->validated();
 
             //Busca por codigo de empleado a los eliminados
-        $operador = Operador::withTrashed()->where('codigo_empleado', $request->input('codigo_empleado'))->first();
+            $operador = Operador::withTrashed()->where('codigo_empleado', $request->input('codigo_empleado'))->first();
 
-          //VALIDACION POR SI EXISTE
-          if ($operador) {
-            if ($operador->trashed()) {
+            //VALIDACION POR SI EXISTE
+            if ($operador) {
+                if ($operador->trashed()) {
+                    return response()->json([
+                        'message' => 'El operador ya existe pero ha sido eliminado. ¿Desea restaurarlo?',
+                        'restore' => true,
+                        'operador_id' => $operador->id
+                    ], 200);
+                }
                 return response()->json([
-                    'message' => 'El operador ya existe pero ha sido eliminado. ¿Desea restaurarlo?',
-                    'restore' => true,
-                    'operador_id' => $operador->id
+                    'message' => 'El concepto ya existe.',
+                    'restore' => false
                 ], 200);
             }
-            return response()->json([
-                'message' => 'El concepto ya existe.',
-                'restore' => false
-            ], 200);
-        }
-        //si no existe el concepto lo crea
-        if(!$operador)
-        {
-            $operador = Operador::create($data);
-            return response(new OperadorResource ($operador), 201);
-        }
+            //si no existe el concepto lo crea
+            if (!$operador) {
+                $operador = Operador::create($data);
+                return response(new OperadorResource($operador), 201);
+            }
 
-        //Lo que ya tenia
-        //$operador = Operador::create($data);
-       // return response(new OperadorResource ($operador), 201);
-        } catch(Exception $e) {
+            //Lo que ya tenia
+            //$operador = Operador::create($data);
+            // return response(new OperadorResource ($operador), 201);
+        } catch (Exception $e) {
             return response()->json([
                 'error' => 'No se pudo guardar el operador'
             ], 500);
         }
-         // me falta el metodo me baso en el concepto controller
+        // me falta el metodo me baso en el concepto controller
         //
     }
 
@@ -90,9 +120,10 @@ class OperadorController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    
+
     public function update(UpdateOperadorRequest $request, string $id)
     {
+        $this->authorize('update', Operador::class);
         //Log::info("id");
         try {
             $data = $request->validated();
@@ -105,8 +136,33 @@ class OperadorController extends Controller
                 'error' => 'No se pudo editar el operador'
             ], 500);
         }
-     
+    }
 
+    public function update_2(UpdateOperadorRequest $request, string $id_user, string $id_operador)
+    {
+        
+            $data = json_decode($request->getContent(), true);
+            $user = User::find($id_user);
+            $user->name = $data["name"];
+            $user->email = $data["email"];
+            if ($data["password"]) {
+                $user->password = bcrypt($data["password"]);
+            }
+            $user->save();
+
+            $operador = Operador::find($id_operador);
+            $operador->id_user = $user->id;
+            $operador->codigo_empleado = $data["codigo_empleado"];
+            $operador->nombre = $data["nombre"];
+            $operador->apellido_paterno = $data["apellido_paterno"];
+            $operador->apellido_materno = $data["apellido_materno"];
+            $operador->CURP = $data["CURP"];
+            $operador->fecha_nacimiento = $data["fecha_nacimiento"];
+            $operador->save();
+
+            return response(new OperadorResource($operador), 201);
+        
+        
     }
 
     /**
@@ -114,10 +170,11 @@ class OperadorController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Operador::class);
         try {
             $operador = Operador::findOrFail($id);
             $operador->delete();
-            return response("Operador eliminado con exito",200);
+            return response("Operador eliminado con exito", 200);
         } catch (ModelNotFoundException $e) {
             return response()->json([
                 'error' => 'No se pudo borrar el operador'
@@ -130,13 +187,12 @@ class OperadorController extends Controller
 
         $operador = Operador::withTrashed()->findOrFail($request->id);
 
-           // Verifica si el registro está eliminado
+        // Verifica si el registro está eliminado
         if ($operador->trashed()) {
 
             // Restaura el registro
             $operador->restore();
             return response()->json(['message' => 'El operador ha sido restaurado.'], 200);
         }
-
     }
 }
